Update install-guide.md
Per https://github.com/tesla-android/tesla-android.github.io/issues/43 I've only had success flashing with Raspberry Pi Imager.  Unless there's a reason not to, I'd suggest only suggesting this one to use.

I also added the suggestion which is in the requirements to update the firmware.  

Lastly it would be nice to know what the recommended SD card size is (and if there are any specific recommendations)